requirements.txt
django>=5.1.6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-storages>=1.13
python-decouple>=3.8
requests>=2.31
Pillow>=10.0
gunicorn>=22.0.0